+++ tests/specs/server-add.js
@@ -83,6 +83,7 @@
     });
 
     describe('server.add-non-autoincrement key', function () {
+        this.timeout(5000);
         var indexedDB = window.indexedDB || window.webkitIndexedDB ||
             window.mozIndexedDB || window.oIndexedDB || window.msIndexedDB;