+++ tests/specs/server-update.js
@@ -169,6 +169,7 @@
     });
 
     describe('server.update-custom-keys', function () {
+        this.timeout(5000);
         var indexedDB = window.indexedDB || window.webkitIndexedDB ||
             window.mozIndexedDB || window.oIndexedDB || window.msIndexedDB;